requirements.txt
Flask>=2.2.3
python-docx==0.8.11
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
pymysql==1.1.0
